+++ src/app/(admin)/noveladmin/page.tsx
@@ -21,8 +21,9 @@ import {
   formCreateProjectSchema,
 } from '@/helpers/Schemas';
 import { transformFormDataNovel } from '@/helpers/TransformFormData';
-import { status, genres, types, nationality } from '@/helpers/Util';
+import { status, types, nationality } from '@/helpers/Util';
 import { createNovel } from '@/hooks/useNovels';
+import { usePublicGenres } from '@/hooks/usePublicApi';
 
 export default function NovelAdmin() {
   const [selectedGenres, setSelectedGenres] = useState<string[]>([]);
@@ -39,6 +40,9 @@ export default function NovelAdmin() {
   } = useForm<InputFormCreateProject>({
     resolver: zodResolver(formCreateProjectSchema),
   });
+
+  const { data: arrayGenres } = usePublicGenres();
+  const genresData = arrayGenres?.data.map((genre) => genre.descricao) || [];
 
   const { mutateAsync: createNovelFn } = useMutation({
     mutationFn: createNovel,
@@ -87,7 +91,10 @@ export default function NovelAdmin() {
     if (response?.statusCode === 400) {
       toaster({
         type: 'error',
-        msg: response.message || 'Erro ao criar a novel',
+        msg:
+          typeof response.message === 'string'
+            ? response.message
+            : response.message?.title || 'Erro ao criar a novel',
       });
       return;
     }
@@ -112,6 +119,7 @@ export default function NovelAdmin() {
           name="cover"
           setValue={setValue}
           errors={errors}
+          defaultValue={''}
         />
 
         <div className="col-span-2 mb-2 flex flex-col gap-4">
@@ -217,7 +225,7 @@ export default function NovelAdmin() {
               getValues={getValues}
               onClick={(key, item) => handleSetMultiValues(key, item)}
               errors={errors}
-              options={genres}
+              options={genresData}
             />
           </div>