+++ src/components/dashboard/table/TableComic.tsx
@@ -179,7 +179,7 @@ export const TableComic = ({ openModal }: ITableComic) => {
     url,
   }: LineTableProps) => {
     const formatedDate = date.replace(/\s\d{2}:\d{2}:\d{2}$/, '');
-    const urlNext = `/project/${url}`;
+    const urlNext = `/project/comic/${url}`;
     return (
       <tr
         onClick={() => router.push(urlNext)}