+++ src/components/dashboard/table/index.tsx
@@ -1,4 +1,4 @@
 export { TableNovel } from './TableNovel';
-export { TableManga } from './TableComic';
+export { TableComic } from './TableComic';
 export { TableStaff } from './TableStaff';
 export { TableActivityStaff } from './TableActivityStaff';