+++ src/hooks/usePrivateApi.ts
@@ -6,9 +6,11 @@ import {
   ApiResponse,
   ErrorResponse,
   IComicResponse,
+  INovelResponse,
   IPrivateComics,
 } from '@/@types/Api';
 import { createComicService, getPrivateComics } from '@/services/ComicService';
+import { getAdminNovelBySlug } from '@/services/NovelService';
 
 export const usePrivateComics = (): UseQueryResult<
   ApiResponse<IPrivateComics>
@@ -23,4 +25,14 @@ export const createComic = async (
   data: FormData,
 ): Promise<IComicResponse | ErrorResponse> => {
   return createComicService(data);
+};
+
+export const useAdminNovelBySlug = (
+  slug: string,
+): UseQueryResult<INovelResponse> => {
+  return useQuery({
+    queryKey: ['adm-novels-slug', slug],
+    queryFn: () => getAdminNovelBySlug(slug),
+    enabled: !!slug,
+  });
 };